pyproject.toml
[project]
name = "hyperlynx"
version = "0.1.0"
description = "Work orders and project management"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "google-genai>=1.49.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "social-auth-app-django>=5.6.0",
]
